requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
fastapi==0.105.0
h11==0.14.0
idna==3.6
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.25.0
